requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
xumm-sdk-py==1.0.4
